Worked on Worker

+++ CSC484project/workers/views.py
@@ -15,7 +15,20 @@ mydb = mysql.connector.connect(
 # Create your views here.
 @login_required
 def worker_page(request):
-    
+    if request.method == 'POST':
+        address = request.POST.get('address2')
+        fName = request.POST.get('fName2')
+        shelter_id = request.POST.get('shelter_id2')
+        lName = request.POST.get('lName2')
+        email = request.POST.get('email2')
+
+
+        cursor = mydb.cursor()
+        cursor.execute("INSERT INTO workers (shelter_id, address, fName, lName, email) VALUES (%s, %s, %s, %s, %s)", (shelter_id, address, fName, lName, email))
+        mydb.commit()
+
+        return redirect('worker_page')
+
     cursor = mydb.cursor(dictionary=True)
     cursor.execute("SELECT * FROM workers")
     workers = cursor.fetchall()
@@ -25,6 +38,11 @@ def worker_page(request):
 @login_required
 def delete_worker(request, worker_id):
     cursor = mydb.cursor()
+    cursor.execute("SELECT * FROM pets WHERE worker_id = %s", (worker_id,))
+    pets = cursor.fetchall()  # Fetch all rows returned by the query
+    if pets:  # Check if there are any related pets
+        return redirect('error_page')
+
     cursor.execute("DELETE FROM workers WHERE worker_id = %s", (worker_id,))
     mydb.commit()
 
@@ -56,6 +74,8 @@ def sort_worker(request):
     shelter_id = request.GET.get('shelter_id')
     fName = request.GET.get('fName')
     lName = request.GET.get('lName')
+    email = request.GET.get('email')
+    worker_id = request.GET.get('worker_id')
     
     
     cursor = mydb.cursor(dictionary=True)
@@ -64,6 +84,10 @@ def sort_worker(request):
     # Adding filters based on provided criteria
     if address:
         query += f" AND address LIKE '%{address}%'"
+    if email:
+        query += f" AND email LIKE '%{email}%'"
+    if worker_id:
+        query += f" AND worker_id LIKE '%{worker_id}%'"
     if shelter_id:
         query += f" AND shelter_id LIKE '%{shelter_id}%'"
     if fName:
